fix: merge rest category infos except for top 3 categories

+++ src/main/java/com/example/dnd6th3moneyroutineserver/expenditure/dto/GoalCategoryInfoDto.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class GoalCategoryInfoDto {
     private CategoryType categoryType;
     private String categoryName;
-    private double percentage;
+    private int percentage;
     private Long expense;
     private List<WeeklyExpenditureDetailDto> weeklyExpenditureDetailDtoList;
 }
